Avanzamento webserver

+++ controller/student_controller.go
@@ -20,15 +20,24 @@ func NewStudentController(studentService service.StudentService) *StudentControl
 }
 
 func (controller *StudentController) Create(writer http.ResponseWriter, requests *http.Request, params httprouter.Params) {
-	studentRequest := request.StudentCreateRequest{}
-	helper.ReadRequestBody(requests, &studentRequest)
-	controller.StudentService.Create(requests.Context(), studentRequest)
-	WebResponse := response.WebResponse{
-		Code:   200,
-		Status: "ok",
-		Data:   nil,
+	valid := service.CheckToken(requests)
+	if valid {
+		studentRequest := request.StudentCreateRequest{}
+		helper.ReadRequestBody(requests, &studentRequest)
+		controller.StudentService.Create(requests.Context(), studentRequest)
+		WebResponse := response.WebResponse{
+			Code:   200,
+			Status: "ok",
+			Data:   nil,
+		}
+		helper.WriteResponse(writer, WebResponse)
+	} else {
+		webRepo := response.WebResponse{
+			Code:   403,
+			Status: "Error token not valid",
+		}
+		helper.WriteResponse(writer, webRepo)
 	}
-	helper.WriteResponse(writer, WebResponse)
 }
 
 func (controller *StudentController) Update(writer http.ResponseWriter, requests *http.Request, params httprouter.Params) {
@@ -48,37 +57,69 @@ func (controller *StudentController) Update(writer http.ResponseWriter, requests
 }
 
 func (controller *StudentController) Delete(writer http.ResponseWriter, requests *http.Request, params httprouter.Params) {
-	studentid := params.ByName("idstudente")
-	id, err := strconv.Atoi(studentid)
-	helper.PanicIfError(err)
-	controller.StudentService.Delete(requests.Context(), id)
-	webRepo := response.WebResponse{
-		Code:   200,
-		Status: "ok",
-		Data:   nil,
+
+	valid := service.CheckToken(requests)
+	if valid {
+		studentid := params.ByName("idstudente")
+		id, err := strconv.Atoi(studentid)
+		helper.PanicIfError(err)
+		controller.StudentService.Delete(requests.Context(), id)
+		webRepo := response.WebResponse{
+			Code:   200,
+			Status: "ok",
+			Data:   nil,
+		}
+		helper.WriteResponse(writer, webRepo)
+	} else {
+		webRepo := response.WebResponse{
+			Code:   403,
+			Status: "Error token not valid",
+		}
+
+		helper.WriteResponse(writer, webRepo)
 	}
-	helper.WriteResponse(writer, webRepo)
 }
 
 func (controller *StudentController) FindAll(writer http.ResponseWriter, requests *http.Request, params httprouter.Params) {
-	result := controller.StudentService.FindAll(requests.Context())
-	webRepo := response.WebResponse{
-		Code:   200,
-		Status: "ok",
-		Data:   result,
+
+	valid := service.CheckToken(requests)
+	if valid {
+		result := controller.StudentService.FindAll(requests.Context())
+		webRepo := response.WebResponse{
+			Code:   200,
+			Status: "ok",
+			Data:   result,
+		}
+		helper.WriteResponse(writer, webRepo)
+	} else {
+		webRepo := response.WebResponse{
+			Code:   403,
+			Status: "Error token not valid",
+		}
+
+		helper.WriteResponse(writer, webRepo)
 	}
-	helper.WriteResponse(writer, webRepo)
 }
 
 func (controller *StudentController) FindById(writer http.ResponseWriter, requests *http.Request, params httprouter.Params) {
-	studentid := params.ByName("idstudente")
-	id, err := strconv.Atoi(studentid)
-	helper.PanicIfError(err)
-	result := controller.StudentService.FindById(requests.Context(), id)
-	webRepo := response.WebResponse{
-		Code:   200,
-		Status: "ok",
-		Data:   result,
+	valid := service.CheckToken(requests)
+	if valid {
+		studentid := params.ByName("idstudente")
+		id, err := strconv.Atoi(studentid)
+		helper.PanicIfError(err)
+		result := controller.StudentService.FindById(requests.Context(), id)
+		webRepo := response.WebResponse{
+			Code:   200,
+			Status: "ok",
+			Data:   result,
+		}
+		helper.WriteResponse(writer, webRepo)
+	} else {
+		webRepo := response.WebResponse{
+			Code:   403,
+			Status: "Error token not valid",
+		}
+
+		helper.WriteResponse(writer, webRepo)
 	}
-	helper.WriteResponse(writer, webRepo)
 }
